# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Finance Management API",
        "description": "Backend API for the personal finance management app.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base URL with port for local development"
        }
    ],
    "paths": {
...
        "/api/v1/categories/": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List all categories",
                "description": "Returns all available transaction categories.\nUsed by the mobile app to populate category pickers.",
                "operationId": "list_categories_api_v1_categories__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/analytics/summary": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Spending summary by category",
                "description": "Returns spending totals grouped by category for the given period.\n\nResponse includes:\n- total_spent       — sum of all categorized transactions\n- category_count    — number of distinct categories used\n- breakdown         — list of {category_id, category_name, total, count, percentage}\n  sorted by total descending\n- unclassified_count — how many transactions still lack a confirmed category\n- period_start / period_end",
                "operationId": "spending_summary_api_v1_analytics_summary_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Period: last_30_days | last_90_days | this_month | last_month",
                            "default": "last_30_days",
                            "title": "Period"
                        },
                        "description": "Period: last_30_days | last_90_days | this_month | last_month"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/trends": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Month-over-month spending trends per category",
                "description": "Returns month-by-month spending totals per category over the last N months.\n\nUseful for the mobile app to render a bar/line chart.\n\nResponse shape:\n{\n    \"months\": [\"2026-01\", \"2026-02\", \"2026-03\"],\n    \"series\": [\n        {\n            \"category_id\":   \"...\",\n            \"category_name\": \"Food & Dining\",\n            \"monthly_totals\": [3200.0, 2800.0, 3100.0]\n        },\n        ...\n    ]\n}",
                "operationId": "spending_trends_api_v1_analytics_trends_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "months",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 12,
                            "minimum": 2,
                            "description": "Number of past months to include",
                            "default": 3,
                            "title": "Months"
                        },
                        "description": "Number of past months to include"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/risk_score": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Financial risk score",
                "operationId": "risk_score_api_v1_analytics_risk_score_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/analytics/spending_summary": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Spending summary (debits only)",
                "operationId": "spending_summary_overview_api_v1_analytics_spending_summary_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Currently only current_month supported",
                            "default": "current_month",
                            "title": "Period"
                        },
                        "description": "Currently only current_month supported"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/monthly_forecast": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Monthly spending forecast",
                "operationId": "monthly_forecast_api_v1_analytics_monthly_forecast_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/analytics/alerts": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Spending alerts",
                "operationId": "spending_alerts_api_v1_analytics_alerts_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/analytics/recommendations": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Financial recommendations",
                "operationId": "recommendations_api_v1_analytics_recommendations_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/analytics/category_breakdown": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Category breakdown (precomputed service)",
                "operationId": "category_breakdown_api_v1_analytics_category_breakdown_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "ISO start date",
                            "title": "Start"
                        },
                        "description": "ISO start date"
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "ISO end date",
                            "title": "End"
                        },
                        "description": "ISO end date"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/monthly_trend": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Monthly trend (precomputed summaries)",
                "operationId": "monthly_trend_api_v1_analytics_monthly_trend_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "months",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 36,
                            "minimum": 1,
                            "default": 3,
                            "title": "Months"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}